Update cron link
The nncron documentation was not exactly right for the node package.  It showed it accepted 6 values but the first value is minutes, the node-cron package uses seconds as the first value.  The official crontab documentation referenced by node-cron only has 5 values, no seconds.  So the best bet I think is to link to node-cron documentation because it points out the differences.