Fix manager refreshing bundle
Newer version of `fs.readdirSync` does not return `.` or `..` and thus excluding the changed bundle from its list. This commit simply makes the temporary file readable by `fs.readdirSync` so that bundle manager can reload it correctly.

Now that we don't have `.` file when reading the bundle folder, I removed the last condition, but there may be a reason for it. Let me know if that is the case.